+++ src/posthog-featureflags.ts
@@ -7,7 +7,6 @@ import {
     EarlyAccessFeatureResponse,
     Properties,
     JsonType,
-    RequestCallback,
 } from './types'
 import { PostHogPersistence } from './posthog-persistence'
 
@@ -193,7 +192,7 @@ export class PostHogFeatureFlags {
             this.instance.requestRouter.endpointFor('api', '/decide/?v=3'),
             { data: encoded_data },
             { method: 'POST' },
-            this.instance._prepare_callback((response) => {
+            (response) => {
                 // reset anon_distinct_id after at least a single request with it
                 // makes it through
                 this.$anon_distinct_id = undefined
@@ -202,7 +201,7 @@ export class PostHogFeatureFlags {
                 // :TRICKY: Reload - start another request if queued!
                 this.setReloadingPaused(false)
                 this._startReloadTimer()
-            }) as RequestCallback
+            }
         )
     }